web-animations/animation-effect/animation-effect-composite.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffect composite attribute test</title>
<meta name="assert" content="The composite operation used to composite this animation effect with the animation stack, as specified by one of the CompositeOperation enumeration values">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationeffect-composite">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
var TIMEOUT = ANIMATION_DURATION / 10;

test(function() {
    var effect = newKeyframeEffect();
    assert_equals(effect.composite, 'replace', 'Wrong default value of ' +
        'composite attribute');
}, 'Test AnimationEffect.composite default value is \'replace\'');

test(function() {
    var effect = newKeyframeEffect({composite: 'accumulate'});
    assert_equals(effect.composite, 'accumulate', 'AnimationEffect.composite should return ' +
        'the value specified in constructor arguments');
}, 'Test AnimationEffect.composite attribute returns value set by constructor');

async_test(function(t) {
    var options = {composite: 'replace'};
    var effect1 = newKeyframeEffect(options);
    var effect2 = newKeyframeEffect(options);
    var target = createDiv(t);

    var animation1 = new Animation(target, effect1, ANIMATION_DURATION);
    var player1 = document.timeline.play(animation1);
    player1.currentTime = ANIMATION_DURATION / 2 - 10;

    var animation2 = new Animation(target, effect2, ANIMATION_DURATION);
    var player2 = document.timeline.play(animation2);
    player2.currentTime = 0;

    player1.ready.then(t.step_func(function() {
        player2.ready.then(t.step_func(function() {
            // player2 is created later than player1
            // so it has greater sequence number and higher priority
            setTimeout(t.step_func(function() {
                assert_approx_equals(window.getComputedStyle(target).top, getExpectedTop(TIMEOUT), 10,
                    'Animation effect should replace the underlying value ' +
                    'in the animation stack');
                t.done();
            }), TIMEOUT + 10);
        }), t.unreached_func('Current ready promise should not be rejected'));
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test AnimationEffect.composite: value \'replace\' forces the effect to replace ' +
    'an underlying value in the animation stack by its intermediate animation value');

async_test(function(t) {
    var options = {composite: 'add'};
    var effect1 = newKeyframeEffect(options);
    var effect2 = newKeyframeEffect(options);
    var target = createDiv(t);

    var animation1 = new Animation(target, effect1, ANIMATION_DURATION);
    var player1 = document.timeline.play(animation1);
    player1.currentTime = ANIMATION_DURATION / 2 - 10;

    var animation2 = new Animation(target, effect2, ANIMATION_DURATION);
    var player2 = document.timeline.play(animation2);
    player2.currentTime = ANIMATION_DURATION / 2 - 10;

    player1.ready.then(t.step_func(function() {
        player2.ready.then(t.step_func(function() {
            setTimeout(t.step_func(function() {
                // Expected top property value is
                // 2 * (Animation top at the moment ANIMATION_DURATION / 2 + 100)
                assert_approx_equals(window.getComputedStyle(target).top,
                    2 * getExpectedTop(ANIMATION_DURATION / 2 + TIMEOUT), 10,
                    'Animation effect should add its intermediate animation value ' +
                    'to an underlying value in the animation stack');
                t.done();
            }), TIMEOUT + 10);
        }), t.unreached_func('Current ready promise should not be rejected'));
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test AnimationEffect.composite: value \'add\' forces the effect to add ' +
    'its intermediate animation value to an underlying value in the animation stack');

// TODO add composite 'accumulate' test
</script>
</body>
